requirements.txt
praw
python-dotenv
openai
fpdf                   
scikit-learn           
textblob               
tqdm                   
streamlit
requests
matplotlib
Pillow



